Fixed corrupt db migration

+++ code/loadBalancer/db/migrate/20131024190945_add_timestamp_to_log_entry.rb
@@ -1,5 +1,5 @@
 class AddTimestampToLogEntry < ActiveRecord::Migration
   def change
-    add_column :log_entries, :timestamp, :long
+    add_column :log_entries, :timestamp, :integer
   end
 end
